1) updating the configurations to run locally; 2) some new test files; 3) new api for search of disabled annotations

+++ annotation-client/src/test/java/eu/europeana/annotation/client/integration/webanno/BaseWebAnnotationDataSetTest.java
@@ -140,6 +140,9 @@ public class BaseWebAnnotationDataSetTest {
 	 */
 	protected void deleteAnnotations(Annotation[] annotations) {
 		WebAnnotationAdminApi webannoAdminApi = new WebAnnotationAdminApiImpl();
+		if(annotations == null) {
+			return;
+		}	
 		for (Annotation annotation : annotations) {
 			ResponseEntity<String> re = webannoAdminApi.deleteAnnotation(
 					annotation.getAnnotationId().getProvider(),
